Adds Next Weather project to portfolio
Adds a new project, Next Weather, to the portfolio.

This project showcases skills in Next.js, Tailwind CSS,
and API integration (OpenWeatherMap). It allows users to
query the current weather and five-day forecast.

Also, increases the gap between elements in the project
card grid and increases the font size of the links